+++ src/app/shared/module/poe/component/item-frame-value/item-frame-value.component.spec.ts
@@ -1,3 +1,4 @@
+import { EventEmitter } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ItemFrameComponent } from '../item-frame/item-frame.component';
 import { ItemFrameValueComponent } from './item-frame-value.component';
@@ -13,6 +14,7 @@ describe('ItemFrameValueComponent', () => {
       providers: [
         {
           provide: ItemFrameComponent, useValue: {
+            queryItemChange: new EventEmitter(),
             onValueChange: () => { },
           }
         }